# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:8000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    x-swagger-router-controller: user
    get:
      description: Get all user of the system
      # used as the method name of the controller
      operationId: getAllUsers
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /adduser:
    x-swagger-router-controller: user
    post:
      parameters:
        - in: body
          name: user
          description: The user to create.
          schema:
            type: object
      operationId: addUser
      responses:
        "200":
          description: Created
        default:
          description: Error

  /deluser/{id}:
    x-swagger-router-controller: user
    parameters:
      - name: id
        in: path
        description: the id of the user to delete
        required: true
        type: string
    delete:
      description: Delete a user of the system
      # used as the method name of the controller
      operationId: delUser
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
            items:
              $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    parameters:
      - name: id
        in: path
        description: the id of the user to retrieve
        required: true
        type: string
    get:
      description: Get a specific user by id
      # used as the method name of the controller
      operationId: getUser
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  User:
    properties:
      id:
        type: number
      gender:
        type: string
      title:
        type: string
      first:
        type: string
      last:
        type: string
      email:
        type: string
      username:
        type: string
      registered:
        type: number
      dob:
        type: number
      phone:
        type: string
      cell:
        type: string
      picture:
        type: string
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string